Updated /api/games route
updated route so that it seeds the database from the nfl url using a findorcreate method.  Basically if it is already there it doesn't create duplicate data.  Added a week and a game status column.  Made it so any games not from the current week have a game status of inactive.